Moved the Slack ping service into a folder as slack-ping
no ref

Relocates the flat single-file service to services/slack-ping, matching
the explore-ping/indexnow-ping naming for outbound ping services and
distinguishing it from the adjacent slack-notifications service (milestone
webhooks). Renames listen() to init() to match the boot lifecycle
convention. Pure move: no behaviour or logic changes.
