+++ ghost/core/core/boot.js
@@ -337,7 +337,7 @@ async function initServices({ghostServer} = {}) {
     const tiers = require('./server/services/tiers');
     const permissions = require('./server/services/permissions');
     const indexnow = require('./server/services/indexnow-ping');
-    const slack = require('./server/services/slack');
+    const slack = require('./server/services/slack-ping');
     const webhooks = require('./server/services/webhooks');
     const postScheduling = require('./server/services/post-scheduling').default;
     const comments = require('./server/services/comments');
@@ -386,7 +386,7 @@ async function initServices({ghostServer} = {}) {
         membersEvents.init(),
         permissions.init(),
         indexnow.init(),
-        slack.listen(),
+        slack.init(),
         audienceFeedback.init(),
         emailService.init({ghostServer}),
         emailAnalytics.init(),
